requirements.txt
appdirs==1.4.4
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
distlib==0.3.2
Django==3.2.8
djangorestframework==3.12.4
filelock==3.0.12
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
virtualenv==20.4.7
Werkzeug==2.0.1
